\name{print.cpf}
\alias{print.cpf}
\title{Print a cpf object}
\description{
  Print method for a \code{cpf} object
}
\usage{
\S3method{print}{cpf}(x, ...)
}
\arguments{
  \item{x}{A \code{cpf} object}
  \item{\dots}{Further arguments to the \code{print} method. Not used
    here.}
}
\value{
  No value returned.
}
\author{Arthur Allignol, \email{arthur.allignol@uni-ulm.de}}
\seealso{\code{\link{cpf}}}
\keyword{print}
